requirements.txt
pyrealsense2==2.56.5.9235
open3d==0.19.0
opencv-python==4.13.0.90
numpy==2.2.6
pyyaml>=6.0
mediapipe>=0.10.0